+++ packages/interledger/src/accounts/types/index.ts
@@ -2,4 +2,3 @@ export * from './accountsService'
 export * from './asset'
 export * from './ilpAccount'
 export * from './ilpBalance'
-export * from './transfer'
