--- packages/interledger/src/accounts/types/transfer.ts
@@ -1,29 +1,0 @@
-export interface Transfer {
-  sourceAccountId: string
-  destinationAccountId: string
-
-  sourceAmount: bigint
-  destinationAmount?: bigint
-}
-
-export interface Transaction {
-  commit: () => Promise<void | TransferError>
-  rollback: () => Promise<void | TransferError>
-}
-
-export enum TransferError {
-  InsufficientBalance = 'InsufficientBalance',
-  InsufficientLiquidity = 'InsufficientLiquidity',
-  InvalidSourceAmount = 'InvalidSourceAmount',
-  InvalidDestinationAmount = 'InvalidDestinationAmount',
-  SameAccounts = 'SameAccounts',
-  TransferAlreadyCommitted = 'TransferAlreadyCommitted',
-  TransferAlreadyRejected = 'TransferAlreadyRejected',
-  TransferExpired = 'TransferExpired',
-  UnknownSourceAccount = 'UnknownSourceAccount',
-  UnknownDestinationAccount = 'UnknownDestinationAccount'
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
-export const isTransferError = (o: any): o is TransferError =>
-  Object.values(TransferError).includes(o)
