+++ packages/interledger/src/connector/core/factories/rafiki-services.ts
@@ -17,9 +17,16 @@ export const RafikiServicesFactory = Factory.define<MockRafikiServices>(
   //  return new InMemoryRouter(peers, { ilpAddress: 'test.rafiki' })
   //})
   .option('ilpAddress', 'test.rafiki')
-  .attr('accounts', ['ilpAddress'], (ilpAddress: string) => {
+  .option('accountService', ['ilpAddress'], (ilpAddress: string) => {
     return new MockAccountsService(ilpAddress)
   })
+  .attr(
+    'accounts',
+    ['accountService'],
+    (accountService: MockAccountsService) => {
+      return accountService
+    }
+  )
   .attr('logger', TestLoggerFactory.build())
   .attr('rates', {
     convert: async (opts) => opts.sourceAmount
@@ -39,4 +46,13 @@ export const RafikiServicesFactory = Factory.define<MockRafikiServices>(
       serverAddress: 'test.rafiki',
       serverSecret: crypto.randomBytes(32)
     })
+  )
+  .attr(
+    'transferService',
+    ['accountService'],
+    (accountService: MockAccountsService) => {
+      return {
+        create: async (transfer) => accountService.transferFunds(transfer)
+      }
+    }
   )
