+++ packages/interledger/src/connector/index.ts
@@ -22,6 +22,7 @@ import {
 } from './core'
 import { Logger } from '../logger/service'
 import { AccountsService } from '../accounts/service'
+import { TransferService } from '../transfer/service'
 
 const ILP_ADDRESS = process.env.ILP_ADDRESS || undefined
 const STREAM_SECRET = process.env.STREAM_SECRET
@@ -33,12 +34,14 @@ interface ServiceDependencies {
   logger?: typeof Logger
   ratesService: RatesService
   accountsService: AccountsService
+  transferService: TransferService
 }
 
 export async function createConnectorService({
   redis,
   ratesService,
-  accountsService
+  accountsService,
+  transferService
 }: ServiceDependencies): Promise<Rafiki> {
   if (!ILP_ADDRESS) {
     throw new Error('ILP_ADDRESS is required')
@@ -72,6 +75,7 @@ export async function createConnectorService({
     accounts: accountsService,
     redis,
     rates: ratesService,
+    transferService,
     stream: {
       serverSecret: STREAM_SECRET,
       serverAddress: ILP_ADDRESS
